Fixed unlink(2) for Java not to remove a directory.

+++ java/src/main/java/jp/gr/java_conf/neko_daisuki/fsyscall/slave/Slave.java
@@ -1969,8 +1969,12 @@ public class Slave implements Runnable {
         catch (UnixException unused) {
             // nothing
         }
+        if (!file.isFile()) {
+            result.setError(file.exists() ? Errno.EISDIR : Errno.ENOENT);
+            return result;
+        }
         if (!file.delete()) {
-            result.setError(Errno.ENOENT);
+            result.setError(Errno.EPERM);
             return result;
         }
